refactor(test): update import paths after lib→domains move (#232)
